refactor: 요금과 할인정책에 대한 인터페이스 삭제
