+++ src/main/java/wooteco/subway/domain/fare/AgeDiscountPolicy.java
@@ -1,12 +1,11 @@
 package wooteco.subway.domain.fare;
 
-public final class AgeDiscountPolicy implements DiscountPolicy {
+public final class AgeDiscountPolicy {
 
     private static final int DEDUCT = 350;
     private static final double TEENAGER_DISCOUNT_RATE = 0.2;
     private static final double CHILD_DISCOUNT_RATE = 0.5;
 
-    @Override
     public Fare discountFare(Fare fare, int age) {
         if (isChild(age)) {
             return fare.discount(DEDUCT, CHILD_DISCOUNT_RATE);
