--- src/main/java/wooteco/subway/domain/fare/DiscountPolicy.java
@@ -1,6 +1,0 @@
-package wooteco.subway.domain.fare;
-
-public interface DiscountPolicy {
-
-    Fare discountFare(Fare fare, int age);
-}
