+++ src/main/java/wooteco/subway/domain/fare/FareCalculator.java
@@ -3,15 +3,15 @@ package wooteco.subway.domain.fare;
 public class FareCalculator {
 
     private final FarePolicy farePolicy;
-    private final DiscountPolicy discountPolicy;
+    private final AgeDiscountPolicy ageDiscountPolicy;
 
-    public FareCalculator(FarePolicy farePolicy, DiscountPolicy discountPolicy) {
+    public FareCalculator(FarePolicy farePolicy, AgeDiscountPolicy ageDiscountPolicy) {
         this.farePolicy = farePolicy;
-        this.discountPolicy = discountPolicy;
+        this.ageDiscountPolicy = ageDiscountPolicy;
     }
 
     public Fare calculate(FareCondition fareCondition) {
         Fare fare = farePolicy.calculateFare(fareCondition.getDistance(), fareCondition.getLine());
-        return discountPolicy.discountFare(fare, fareCondition.getAge());
+        return ageDiscountPolicy.discountFare(fare, fareCondition.getAge());
     }
 }
