+++ src/test/java/wooteco/subway/domain/fare/FareCalculatorTest.java
@@ -12,7 +12,7 @@ class FareCalculatorTest {
     @Test
     @DisplayName("조건이 주어졌을 때 요금을 계산한다.")
     void calculate() {
-        FareCalculator fareCalculator = new FareCalculator(new FarePolicyImpl(), new AgeDiscountPolicy());
+        FareCalculator fareCalculator = new FareCalculator(new FarePolicy(), new AgeDiscountPolicy());
 
         Line line1 = new Line("2호선", "green", 500);
         Line line2 = new Line("신분당선", "red", 1000);
